Moved buildRoutine to lib
